pyproject.toml
[project]
name = "core"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "wagtail (>=6.3.2,<7.0.0)",
    "wagtail-localize (>=1.11.2,<2.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.6"

[tool.ruff.lint]
select = [
  "A",      # flake8-builtins
  "ANN",    # flake8-annotations
  "ARG",    # flake8-unused-arguments
  "ASYNC",  # flake8-async
  "B",      # flake8-bugbear
  "C4",     # flake8-comprehensions
  "C90",    # mccabe
  "DTZ",    # flake8-datetimez
  "E",      # pycodestyl
  "EM",     # flake8-errmsg
  "ERA",    # eradicate
  "F",      # pyflakes
  "FBT",    # flake8-boolean-trap
  "I",      # isort
  "INT",    # flake8-gettext
  "N",      # pep8-naming
  "PL",     # Pylint
  "PERF",   # Perflint
  "PT",     # flake8-pytest-style
  "RET",    # flake8-return
  "RUF",    # Ruff-specific rules
  "S",      # flake8-bandit
  "SLF",    # flake8-flake8-self
  "T20",    # flake8-print
  "TRY",    # tryceratops
  "SIM",    # flake8-simplify
  "UP",     # pyupgrade
  "W",      # pycodestyle
]
ignore = ["ANN101", "RUF001", "RUF002", "RUF003", "RET504"]
ignore-init-module-imports = true

[tool.ruff.lint.extend-per-file-ignores]
"**/{tests,docs,tools}/*" = ["E402", "ANN401", "ANN001", "S603", "ARG001", "S101", "S311"]

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"

[tool.ruff.lint.isort]
lines-between-types = 1
lines-after-imports = 2

[tool.ruff.lint.mccabe]
# Flag errors (C901) whenever the complexity level exceeds 6.
max-complexity = 6
